Accept video url as input

+++ main_ui.py
@@ -7,12 +7,49 @@ from pydub import AudioSegment
 import urllib.parse
 import requests
 from gtts import gTTS
-import shutil 
+import shutil
+import yt_dlp
+import io
 
 temp_folder = "resources"
 os.makedirs(temp_folder, exist_ok=True)
 # File uploader for MP4 video
 video_file = st.file_uploader("Upload an MP4 video file", type=["mp4"])
+video_url = st.text_input("Enter the URL of an MP4 video")
+if video_url:
+    if "youtube.com" in video_url or "youtu.be" in video_url:
+        # Use pytube for YouTube links
+        with st.spinner("Downloading video from URL..."):
+            try:
+                # Use yt-dlp for YouTube links
+                ydl_opts = {
+                    'format': 'bestvideo+bestaudio/best',
+                    'outtmpl': os.path.join(temp_folder, "uploaded_video.mp4"),  # Output to stdout
+                    'merge_output_format': 'mp4',  # Merge video and audio into mp4
+                    'quiet': False,  # Suppress output
+                }
+                with yt_dlp.YoutubeDL(ydl_opts) as ydl:
+                    info = ydl.extract_info(video_url, download=True)
+                    # ydl.download([video_url])
+                    video_file = ydl.prepare_filename(info).replace(".webm", ".mp4").replace(".mkv", ".mp4") #os.path.join(temp_folder, "uploaded_video.mp4")
+                st.success("YouTube video downloaded successfully! "+ydl.extract_info(video_url, download=False)['title'])
+            except Exception as e:
+                st.error(f"An error occurred: {e}")
+    else:
+        try:
+            with st.spinner("Downloading video from URL..."):
+                response = requests.get(video_url, stream=True)
+                if response.status_code == 200:
+                    video_file = response.raw
+                    # uploaded_video_path = os.path.join(temp_folder, "downloaded_video.mp4")
+                    # with open(uploaded_video_path, "wb") as f:
+                    #     shutil.copyfileobj(response.raw, f)
+                    # st.success("Video downloaded successfully!")
+                    # st.video(uploaded_video_path, muted=False)
+                else:
+                    st.error("Failed to download video. Please check the URL.")
+        except Exception as e:
+            st.error(f"An error occurred: {e}")
 
 
 def extract_audio_from_video(video_path):
@@ -32,7 +69,7 @@ def separate_audio_layers(audio_path):
 
 def transcribe_audio(audio_path):
     with st.spinner("🧠 Transcribing audio..."):
-        model = whisper.load_model("medium")
+        model = whisper.load_model("large")
         result = model.transcribe(
             audio_path, language="ja", word_timestamps=False, fp16=False)
         print(result["segments"])
@@ -75,7 +112,8 @@ def google_translate_segments(segments, source_lang="ja", target_lang="en"):
 
             # Translate
             try:
-                translation = google_translate_text(original_text, source_lang, target_lang)
+                translation = google_translate_text(
+                    original_text, source_lang, target_lang)
             except Exception as e:
                 print(f"Translation failed for segment {idx}: {e}")
                 translation = "..."
@@ -95,18 +133,14 @@ def google_translate_segments(segments, source_lang="ja", target_lang="en"):
             if len(spoken) > duration_ms:
                 spoken = spoken[:duration_ms]
             else:
-                spoken += AudioSegment.silent(duration=duration_ms - len(spoken))
+                spoken += AudioSegment.silent(
+                    duration=duration_ms - len(spoken))
 
             # Add processed speech
             final_audio += spoken
             last_end_time = segment["end"]
-            # if len(spoken) > duration_ms:
-            #     spoken = spoken[:duration_ms]
-            # else:
-            #     spoken += AudioSegment.silent(duration=(duration_ms - len(spoken)))
-            # final_audio += spoken
-
-            print(f"✅ Segment {idx+1}/{len(segments)} | ${segment["start"]} {translation}")
+            print(
+                f"✅ Segment {idx+1}/{len(segments)} | ${segment["start"]} {translation}")
 
             # 5. Save final audio
         final_audio.export(translated_audio, format="wav")
@@ -118,8 +152,9 @@ def google_translate_segments(segments, source_lang="ja", target_lang="en"):
 if video_file is not None:
     st.video(video_file, muted=False)
     uploaded_video_path = os.path.join(temp_folder, "uploaded_video.mp4")
-    with open(uploaded_video_path, "wb") as f:
-        f.write(video_file.read())
+    if not os.path.exists(uploaded_video_path):
+        with open(uploaded_video_path, "wb") as f:
+            f.write(video_file.read())
 
     # Extract audio from the video
     audio_path = extract_audio_from_video(uploaded_video_path)
@@ -153,7 +188,8 @@ if video_file is not None:
 
     segments = transcribe_audio(vocals_path)
     translated_audio = google_translate_segments(segments)
-    st.write("🎤 Translated audio: "); st.audio(translated_audio)
+    st.write("🎤 Translated audio: ")
+    st.audio(translated_audio)
 
     # Combine audio tracks
     video = VideoFileClip(uploaded_video_path)
@@ -163,11 +199,13 @@ if video_file is not None:
     drums_audio = AudioFileClip(drums_path)
     bass_audio = AudioFileClip(bass_path)
     # Combine audio tracks
-    combined_audio = CompositeAudioClip([translated_audio, bg_audio, drums_audio, bass_audio])
+    combined_audio = CompositeAudioClip(
+        [translated_audio, bg_audio, drums_audio, bass_audio])
 
     video = video.with_audio(combined_audio)
     dubbed_video_path = os.path.join(temp_folder, "dubbed_video.mp4")
-    video.write_videofile(dubbed_video_path, codec="libx264", audio_codec="aac")
+    video.write_videofile(
+        dubbed_video_path, codec="libx264", audio_codec="aac")
     st.success("🎥 Dubbed video saved")
     st.video(dubbed_video_path)
     # Clean up temporary files
